refactor: rename calendar/util to calendarHelper for consistency and so we dont have two util files
