+++ src/content/Calendar/Calendar.tsx
@@ -1,5 +1,5 @@
 import { ISectionData, Term } from "../App/App.types"
-import { convertToMatrix, getEndHour } from "./utils"
+import { convertToMatrix, getEndHour } from "./calendarHelper"
 import "./Calendar.css"
 import { useContext } from "react"
 import { ModalDispatchContext, ModalPreset } from "../ModalLayer"
